+++ docxbox/app/app.cc
@@ -102,7 +102,9 @@ bool App::Process() {
 
   // Process
   switch (command) {
-    case AppCommands::Command_FileDiff:return docx_archive->ViewFilesDiff("");
+    case AppCommands::Command_ExecuteUserCommand:
+      return docx_archive->ExecuteUserCommand();
+    case AppCommands::Command_FileDiff:return docx_archive->ViewFilesDiff();
     case AppCommands::Command_Help: {
       AppCommands::Commands kCommand;
       std::string command_identifier;